requirements.txt
# Core Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
python-dotenv>=1.0.0

# LangChain & LangGraph
langgraph>=1.1.0
langchain>=1.2.0
langchain-community>=0.4.0
langchain-core>=1.3.0
langchain-openai>=0.0.2
langchain-anthropic>=0.0.1
langchain-huggingface>=0.0.1

# LLM APIs
openai>=1.6.1
anthropic>=0.8.1

# MCP Server
mcp>=0.1.0

# OCR & Document Processing
paddleocr>=2.7.0
paddlepaddle>=2.5.0
opencv-python>=4.8.0
pymupdf>=1.23.0
python-docx>=1.1.0
openpyxl>=3.1.0
pandas>=2.1.0

# Speech Recognition
paddlespeech>=1.4.0
funasr>=1.0.0
modelscope>=1.20.0

# Computer Vision & Behavior Analysis
ultralytics>=8.0.0

# RAG & Embeddings
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
tiktoken>=0.5.0
tavily-python>=0.3.0

# Web Scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
aiohttp>=3.9.0

# Utilities
numpy>=1.24.0
pytz>=2023.3
uuid>=1.30
json5>=2.0.1

# Logging & Monitoring
loguru>=0.7.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0